lan_gui/requirements.txt
mss==9.0.1
opencv-python==4.8.0.76
requests==2.31.0
Pillow==10.0.
keyboard
PyQt5
json_numpy
